requirements.txt
flask
gunicorn
numpy
pandas
sklearn
PyYAML
evidently
dill
six
matplotlib
-e .